+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * new argument `round` to `st_coordinates` to round output coordinates to a given precision level; #2620
 
+* `st_combine()` preserves the precision of input object; #2618
+
 # version 1.1-2
 
 * remove use of `.Dim` in `structure()`